% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap_base.R
\name{create_tmap_base}
\alias{create_tmap_base}
\title{Create tmap base map object}
\usage{
create_tmap_base(
  style = "atlas",
  type = "base_full",
  details = TRUE,
  lines = TRUE,
  labels = TRUE
)
}
\arguments{
\item{style}{The style of the base map to use must be:
\itemize{
\item "atlas (the default): and RPG color base map
\item "light": a clean, simple light base map
\item "dark": a clean, simple dark base map
}}

\item{type}{The type of base map to create must be:
\itemize{
\item "base_full" (the default): creates base map to go under all mapping,
includes labels and lines
\item "base_detailed": creates a detailed base including land use and land cover
\item "base_simple": creates a clean base with hill shade only
}}

\item{details}{A Boolean.
\itemize{
\item \code{TRUE} (the default): ensures that details will be added to non full bases
\item \code{FALSE}: ensures that no details will be added to non full bases
}}

\item{lines}{A Boolean.
\itemize{
\item \code{TRUE} (the default): ensures that lines will be added to non full bases
\item \code{FALSE}: ensures that no lines will be added to non full bases
}}

\item{labels}{A Boolean.
\itemize{
\item \code{TRUE} (the default): ensures that labels will be added to non full bases
\item \code{FALSE}: ensures that no labels will be added to non full bases
}}
}
\value{
A tmap object
}
\description{
Creates a base map tmap object that can be passed to further mapping with
the appropriate z index for lines and labels if used.
}
